chore(Mixpanel): move sdk loader to different file
